LayoutTests/fast/mediastream/RTCPeerConnection-have-local-pranswer-expected.txt

Tests RTCPeerConnection in have-local-pranswer state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.signalingState is "stable"
PASS pc.setRemoteDescription(sessionDescription, requestSucceeded1, finishIfFailed); did not throw exception.
FAIL setRemoteDescription failed.
PASS successfullyParsed is true

TEST COMPLETE
